Update all packages after meta-package building.
If multiple packages were built at once, the parallel nature of
the functions meant that the repo DB file might be out of sync
and overwritten. To overcome that, the meta-package (built last)
message contains a list of all completed built packages, which
are pulled from s3 and re-added to the DB file.